Minor test repackaging
